# frontend/app/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/quicklooks/{id}/tiles/{z}/{y}/{x}": {
            "get": {
                "summary": "Get Tile",
                "operationId": "get_tile",
                "parameters": [
                    {
                        "name": "z",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Z"
                        }
                    },
                    {
                        "name": "y",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Y"
                        }
                    },
                    {
                        "name": "x",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "X"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "^\\w+:\\w+$",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/quicklooks/{id}/fits_header/{ccd_name}": {
            "get": {
                "summary": "Get Fits Header",
                "operationId": "get_fits_header",
                "parameters": [
                    {
                        "name": "ccd_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Ccd Name"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "^\\w+:\\w+$",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HeaderType"
                                    },
                                    "title": "Response Get Fits Header Api Quicklooks  Id  Fits Header  Ccd Name  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/quicklooks": {
            "get": {
                "summary": "List Quicklooks",
                "operationId": "list_quicklooks",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/QuicklookStatus"
                                    },
                                    "type": "array",
                                    "title": "Response List Quicklooks Api Quicklooks Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Quicklook",
                "description": "Create a quicklook",
                "operationId": "create_quicklook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuicklookCreateFrontend"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CardType": {
                "prefixItems": [
                    {
                        "type": "string"
                    },
                    {
                        "type": "string"
                    },
                    {
                        "type": "string"
                    },
                    {
                        "type": "string"
                    }
                ],
                "type": "array",
                "maxItems": 4,
                "minItems": 4
            },
...
            "GenerateProgress": {
                "properties": {
                    "download": {
                        "$ref": "#/components/schemas/Progress"
                    },
                    "preprocess": {
                        "$ref": "#/components/schemas/Progress"
                    },
                    "maketile": {
                        "$ref": "#/components/schemas/Progress"
                    }
                },
                "type": "object",
                "required": [
                    "download",
                    "preprocess",
                    "maketile"
                ],
                "title": "GenerateProgress"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "HeaderType": {
                "items": {
                    "$ref": "#/components/schemas/CardType"
                },
                "type": "array"
            },
...
            "MergeProgress": {
                "properties": {
                    "merge": {
                        "$ref": "#/components/schemas/Progress"
                    }
                },
                "type": "object",
                "required": [
                    "merge"
                ],
                "title": "MergeProgress"
            },
...
            "Progress": {
                "properties": {
                    "count": {
                        "type": "integer",
                        "title": "Count"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    }
                },
                "type": "object",
                "required": [
                    "count",
                    "total"
                ],
                "title": "Progress"
            },
            "QuicklookCreateFrontend": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "id"
                ],
                "title": "QuicklookCreateFrontend"
            },
            "QuicklookJobPhase": {
                "type": "integer",
                "enum": [
                    0,
                    1,
                    2,
                    3,
                    4,
                    5,
                    6,
                    7,
                    -1
                ],
                "title": "QuicklookJobPhase"
            },
...
            "QuicklookStatus": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "phase": {
                        "$ref": "#/components/schemas/QuicklookJobPhase"
                    },
                    "generate_progress": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "$ref": "#/components/schemas/GenerateProgress"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Generate Progress"
                    },
                    "transfer_progress": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "$ref": "#/components/schemas/TransferProgress"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Transfer Progress"
                    },
                    "merge_progress": {
                        "anyOf": [
                            {
                                "additionalProperties": {
                                    "$ref": "#/components/schemas/MergeProgress"
                                },
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Merge Progress"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "phase",
                    "generate_progress",
                    "transfer_progress",
                    "merge_progress"
                ],
                "title": "QuicklookStatus"
            },
...
            "TransferProgress": {
                "properties": {
                    "transfer": {
                        "$ref": "#/components/schemas/Progress"
                    }
                },
                "type": "object",
                "required": [
                    "transfer"
                ],
                "title": "TransferProgress"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}